kotlin-maven-plugin-tools/src/main/kotlin/com/github/gantsign/maven/tools/plugin/extractor/kotlin/internal/kotlinc/MavenMessageCollector.kt
/*-
 * #%L
 * kotlin-maven-plugin-tools
 * %%
 * Copyright (C) 2018 - 2022 GantSign Ltd.
 * %%
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *      http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 * #L%
 */
package com.github.gantsign.maven.tools.plugin.extractor.kotlin.internal.kotlinc

import org.codehaus.plexus.logging.Logger
import org.jetbrains.kotlin.cli.common.messages.CompilerMessageSeverity
import org.jetbrains.kotlin.cli.common.messages.CompilerMessageSourceLocation
import org.jetbrains.kotlin.cli.common.messages.MessageCollector
import org.jetbrains.kotlin.cli.common.messages.MessageRenderer

internal class MavenMessageCollector(
    private val logger: Logger,
    private val messageRenderer: MessageRenderer
) : MessageCollector {
    private var hasErrors = false

    override fun clear() {
        hasErrors = false
    }

    override fun report(
        severity: CompilerMessageSeverity,
        message: String,
        location: CompilerMessageSourceLocation?
    ) {
        if (CompilerMessageSeverity.VERBOSE.contains(severity)) return
        hasErrors = hasErrors or severity.isError
        logger.warn(messageRenderer.render(severity, message, location))
    }

    override fun hasErrors(): Boolean = hasErrors
}
